requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
jinja2==3.1.5
python-dotenv==1.0.1
python-multipart==0.0.20
greenlet==3.2.3
